+++ graphql-java-client-runtime/src/main/java/com/graphql_java_generator/client/graphqlrepository/GraphQLRepositoryInvocationHandler.java
@@ -364,7 +364,7 @@ public class GraphQLRepositoryInvocationHandler<T> implements InvocationHandler 
 			registeredMethod.executorMethod = registeredMethod.executor.getClass()
 					.getMethod(registeredMethod.executorMethodName, registeredMethod.executorParameterTypes);
 		} catch (NoSuchMethodException | SecurityException e) {
-			StringBuffer parameters = new StringBuffer();
+			StringBuilder parameters = new StringBuilder();
 			String separator = "";
 			for (Class<?> clazz : registeredMethod.executorParameterTypes) {
 				parameters.append(separator);