+++ graphql-maven-plugin-logic/src/main/java/com/graphql_java_generator/plugin/language/Field.java
@@ -186,6 +186,17 @@ public interface Field {
 	public Value<?> getDefaultValue();
 
 	/**
+	 * Returns the default value, as text, as it can be written into a generated GraphQL schema.<BR/>
+	 * A <I>str</I> string default value will be returned as <I>"str"</I>,a <I>JEDI</I> enum value will be returned as
+	 * <I>JEDI</I>, ...
+	 * 
+	 * @return
+	 */
+	default public String getDefaultValueAsText() {
+		return GraphqlUtils.graphqlUtils.getValueAsText(getDefaultValue());
+	}
+
+	/**
 	 * Returns the {@link Relation} description for this field.
 	 * 
 	 * @return null if this field is not a relation to another Entity
@@ -201,8 +212,11 @@ public interface Field {
 	 */
 	public String getAnnotation();
 
-	/** Returns the comments that have been found for this object, in the provided GraphQL schema */
+	/** Returns the comments that have been found before this object, in the provided GraphQL schema */
 	public List<String> getComments();
+
+	/** Returns the description for this object, in the provided GraphQL schema */
+	public Description getDescription();
 
 	/**
 	 * Convert the given name, which is supposed to be in camel case (for instance: thisIsCamelCase) to a pascal case