+++ graphql-maven-plugin-logic/src/main/java/com/graphql_java_generator/plugin/language/impl/DirectiveImpl.java
@@ -6,6 +6,7 @@ package com.graphql_java_generator.plugin.language.impl;
 import java.util.ArrayList;
 import java.util.List;
 
+import com.graphql_java_generator.plugin.language.Description;
 import com.graphql_java_generator.plugin.language.Directive;
 import com.graphql_java_generator.plugin.language.DirectiveLocation;
 import com.graphql_java_generator.plugin.language.Field;
@@ -29,8 +30,11 @@ public class DirectiveImpl implements Directive {
 	/** Returns the list of location that this directive may have */
 	private List<DirectiveLocation> directiveLocations = new ArrayList<>();
 
-	/** The comments that have been found for this object, in the provided GraphQL schema */
+	/** The comments that have been found before this object, in the provided GraphQL schema */
 	private List<String> comments = new ArrayList<>();
+
+	/** The description of this directive, in the provided GraphQL schema */
+	private Description description = null;
 
 	/**
 	 * True if this directive is a standard GraphQL directive, or if it has been defined in the GraphQL schema. Default