+++ graphql-maven-plugin-logic/src/main/resources/templates/client_GraphQLRequest.vm.java
@@ -405,7 +405,7 @@ public class GraphQLRequest${springBeanSuffix} extends ObjectResponse {
 			if (parameters == null) {
 				logger.trace("Executing of " + requestType.toString() + " without parameters");
 			} else {
-				StringBuffer sb = new StringBuffer("Executing of root mutation with parameters: ");
+				StringBuilder sb = new StringBuilder("Executing of root mutation with parameters: ");
 				boolean addComma = false;
 				for (String key : parameters.keySet()) {
 					sb.append(key).append(":").append(parameters.get(key));