+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/impl/DataFetchersDelegateReservedJavaKeywordAllFieldCasesImpl.java
@@ -1,0 +1,67 @@
+package org.allGraphQLCases.server.impl;
+
+import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
+
+import javax.annotation.Resource;
+
+import org.allGraphQLCases.server.SIP_WithID_SIS;
+import org.allGraphQLCases.server.STP_Human_STS;
+import org.allGraphQLCases.server.STP_ReservedJavaKeywordAllFieldCases_STS;
+import org.allGraphQLCases.server.SUP_AnyCharacter_SUS;
+import org.allGraphQLCases.server.util.DataFetchersDelegateReservedJavaKeywordAllFieldCases;
+import org.dataloader.DataLoader;
+import org.springframework.stereotype.Component;
+
+import graphql.schema.DataFetchingEnvironment;
+
+@Component
+public class DataFetchersDelegateReservedJavaKeywordAllFieldCasesImpl
+		implements DataFetchersDelegateReservedJavaKeywordAllFieldCases {
+
+	@Resource
+	DataGenerator generator;
+
+	@Override
+	public CompletableFuture<SIP_WithID_SIS> _implements(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, SIP_WithID_SIS> dataLoader, STP_ReservedJavaKeywordAllFieldCases_STS origin) {
+		return CompletableFuture.completedFuture(generator.generateInstance(SIP_WithID_SIS.class));
+	}
+
+	@Override
+	public SIP_WithID_SIS _implements(DataFetchingEnvironment dataFetchingEnvironment,
+			STP_ReservedJavaKeywordAllFieldCases_STS origin) {
+		return generator.generateInstance(SIP_WithID_SIS.class);
+	}
+
+	@Override
+	public CompletableFuture<STP_Human_STS> _int(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, STP_Human_STS> dataLoader, STP_ReservedJavaKeywordAllFieldCases_STS origin) {
+		return CompletableFuture.completedFuture(generator.generateInstance(STP_Human_STS.class));
+	}
+
+	@Override
+	public STP_Human_STS _int(DataFetchingEnvironment dataFetchingEnvironment,
+			STP_ReservedJavaKeywordAllFieldCases_STS origin) {
+		return generator.generateInstance(STP_Human_STS.class);
+	}
+
+	@Override
+	public SUP_AnyCharacter_SUS _interface(DataFetchingEnvironment dataFetchingEnvironment,
+			STP_ReservedJavaKeywordAllFieldCases_STS origin) {
+		return generator.generateInstance(SUP_AnyCharacter_SUS.class);
+	}
+
+	@Override
+	public CompletableFuture<SIP_WithID_SIS> nonJavaKeywordField(DataFetchingEnvironment dataFetchingEnvironment,
+			DataLoader<UUID, SIP_WithID_SIS> dataLoader, STP_ReservedJavaKeywordAllFieldCases_STS origin) {
+		return CompletableFuture.completedFuture(generator.generateInstance(SIP_WithID_SIS.class));
+	}
+
+	@Override
+	public SIP_WithID_SIS nonJavaKeywordField(DataFetchingEnvironment dataFetchingEnvironment,
+			STP_ReservedJavaKeywordAllFieldCases_STS origin) {
+		return generator.generateInstance(SIP_WithID_SIS.class);
+	}
+
+}